[DO NOT MERGE] Bare Auth Plugin PoC
This is a proof-of-concept of a new auth plugin that is *not* built on
the AWS Mobile SDK.

Instead, this plugin calls Cognito Identity Provider directly through a
low-level code-generated client.

The PoC shows sign-up, confirmation with email code, and sign-in with
SRP_A. Aside from the Cognito interactions needed to fulfill SRP_A, this
PoC *does not* include support for MFA on sign-up.

A few findings from this exercise:

1. It's not that bad, really.
2. SRP and hashing are hairy implementation details
3. The system works like a state machine.

The realization of point (3) is the biggest.

Either:

1. We need an architecture to accomodate a variety of interrelated
flows, that can recycle implementation components, or

2. We should just make a lot of different plugins. It might be easier
for the customer if they can just install a different plugin for
different auth flows.